Architecture: Updated folder structure for easy readability of the code.
