+++ app/src/main/AndroidManifest.xml
@@ -10,7 +10,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.MpesaAppUIClone">
         <activity
-            android:name="com.breens.mpesaappuiclone.MainActivity"
+            android:name="com.breens.mpesaappuiclone.presentation.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
