+++ app/src/main/java/com/breens/mpesaappuiclone/data/Statement.kt
@@ -1,4 +1,4 @@
-package com.breens.mpesaappuiclone
+package com.breens.mpesaappuiclone.data
 
 data class Statement(
     val initials: String,
