+++ app/src/main/java/com/breens/mpesaappuiclone/presentation/StatementAdapter.kt
@@ -1,4 +1,4 @@
-package com.breens.mpesaappuiclone
+package com.breens.mpesaappuiclone.presentation
 
 
 import android.view.LayoutInflater
@@ -6,6 +6,8 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
+import com.breens.mpesaappuiclone.R
+import com.breens.mpesaappuiclone.data.Statement
 import io.getstream.avatarview.AvatarView
 
 class StatementAdapter(private val mList: List<Statement>) : RecyclerView.Adapter<StatementAdapter.ViewHolder>() {
