+++ app/src/main/res/layout/fragment_discover.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FragmentDiscover">
+    tools:context=".presentation.FragmentDiscover">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
