+++ app/src/main/res/layout/fragment_home.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="15dp"
-    tools:context=".FragmentHome">
+    tools:context=".presentation.FragmentHome">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/appBarLayout"
